Factor out shareable methods to GeocodableSync (temporary name)
